requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
beautifulsoup4==4.12.0
requests==2.31.0
sentence-transformers==2.2.2
numpy==1.24.0
scikit-learn==1.3.0